Refactor Database utils
