+++ src/k8s/pkg/k8sd/app/hooks_join.go
@@ -2,7 +2,7 @@ package app
 
 import (
 	"fmt"
-	databaseutil "github.com/canonical/k8s/pkg/k8sd/database/utils"
+	databaseutil "github.com/canonical/k8s/pkg/k8sd/database/util"
 	"net"
 
 	apiv1 "github.com/canonical/k8s/api/v1"
