+++ src/k8s/pkg/k8sd/database/util/kubernetes_auth_tokens.go
@@ -1,4 +1,4 @@
-package utils
+package databaseutil
 
 import (
 	"context"
