+++ src/k8s/pkg/k8sd/database/util/node.go
@@ -1,32 +1,12 @@
-package utils
+package databaseutil
 
 import (
 	"context"
 	"database/sql"
 	"fmt"
-
 	"github.com/canonical/k8s/pkg/k8sd/database"
-	"github.com/canonical/k8s/pkg/k8sd/types"
 	"github.com/canonical/microcluster/state"
 )
-
-// GetClusterConfig is a convenience wrapper around the database call to get the cluster config.
-func GetClusterConfig(ctx context.Context, state *state.State) (types.ClusterConfig, error) {
-	var clusterConfig types.ClusterConfig
-	var err error
-
-	if err := state.Database.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-		clusterConfig, err = database.GetClusterConfig(ctx, tx)
-		if err != nil {
-			return fmt.Errorf("failed to get cluster config from database: %w", err)
-		}
-		return nil
-	}); err != nil {
-		return types.ClusterConfig{}, fmt.Errorf("failed to perform cluster config transaction request: %w", err)
-	}
-
-	return clusterConfig, nil
-}
 
 // CheckWorkerExists is a convenience wrapper around the database call to check if a worker node entry exists.
 func CheckWorkerExists(ctx context.Context, state *state.State, name string) (bool, error) {
@@ -61,4 +41,13 @@ func DeleteWorkerNodeEntry(ctx context.Context, state *state.State, name string)
 	}
 
 	return nil
+}
+
+// IsWorkerNode returns true if the given node name belongs to a worker node in the cluster.
+func IsWorkerNode(ctx context.Context, s *state.State, name string) (bool, error) {
+	exists, err := CheckWorkerExists(ctx, s, name)
+	if err != nil {
+		return false, fmt.Errorf("failed to check if worker node %q exists: %w", name, err)
+	}
+	return exists, nil
 }
