+++ src/k8s/pkg/utils/node.go
@@ -3,8 +3,6 @@ package utils
 import (
 	"context"
 	"fmt"
-	"github.com/canonical/k8s/pkg/k8sd/database/utils"
-
 	apiv1 "github.com/canonical/k8s/api/v1"
 	"github.com/canonical/microcluster/state"
 )
@@ -42,15 +40,6 @@ func IsControlPlaneNode(ctx context.Context, s *state.State, name string) (bool,
 		return false, fmt.Errorf("failed to get control-plane node: %w", err)
 	}
 	return node != nil, nil
-}
-
-// IsWorkerNode returns true if the given node name belongs to a worker node in the cluster.
-func IsWorkerNode(ctx context.Context, s *state.State, name string) (bool, error) {
-	exists, err := utils.CheckWorkerExists(ctx, s, name)
-	if err != nil {
-		return false, fmt.Errorf("failed to check if worker node %q exists: %w", name, err)
-	}
-	return exists, nil
 }
 
 // DatastoreRoleFromString converts the string-based role to the enum-based role.
